Рефакторинг: вынос логики выбора типа storage в server
