+++ internal/server/server.go
@@ -7,9 +7,20 @@ import (
 	"github.com/k0st1a/metrics/internal/handlers"
 	"github.com/k0st1a/metrics/internal/handlers/json"
 	"github.com/k0st1a/metrics/internal/handlers/text"
-	"github.com/k0st1a/metrics/internal/storage"
+	"github.com/k0st1a/metrics/internal/storage/file"
+	"github.com/k0st1a/metrics/internal/storage/inmemory"
 	"github.com/rs/zerolog/log"
 )
+
+type Storage interface {
+	GetGauge(string) (float64, bool)
+	StoreGauge(string, float64)
+
+	GetCounter(string) (int64, bool)
+	StoreCounter(string, int64)
+
+	GetAll() (map[string]int64, map[string]float64)
+}
 
 func Run() error {
 	log.Debug().Msg("Run server")
@@ -21,12 +32,19 @@ func Run() error {
 
 	printConfig(cfg)
 
-	r := handlers.NewRouter()
+	var s Storage
+	if cfg.FileStoragePath == "" {
+		log.Debug().Msg("Using memory storage")
+		s = inmemory.NewStorage()
+	} else {
+		log.Debug().Msg("Using file storage")
+		s = file.NewStorage(cfg.FileStoragePath, cfg.StoreInterval, cfg.Restore)
+	}
 
-	s := storage.NewStorage(cfg.FileStoragePath, cfg.StoreInterval, cfg.Restore)
 	th := text.NewHandler(s)
 	jh := json.NewHandler(s)
 
+	r := handlers.NewRouter()
 	text.BuildRouter(r, th)
 	json.BuildRouter(r, jh)
